requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-huggingface>=0.0.3
sentence-transformers>=2.7.0
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.13.0
python-dotenv>=1.0.0
torch>=2.0.0
